Refactor for Accountability
