+++ src/components/index.js
@@ -4,6 +4,7 @@ export { default as Button } from "./Button";
 export { default as HamburgerButton } from "./HamburgerButton";
 export { default as HoverImage } from "./HoverImage";
 export { default as Input } from "./Input";
+export { default as Icon } from "./Icon";
 export { default as Label } from "./Label";
 export { default as Link } from "./Link";
 export { default as Navbar } from "./Navbar";
