+++ src/views/Home/Accountability/TitleContainer/index.js
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+const TitleContainer = styled.div``;
+
+export default TitleContainer;
